refactor: moved user login details to models
